refactor: Main Header 분리 및 Router Switch 적용
