+++ frontend/src/App.tsx
@@ -1,31 +1,21 @@
-import { FormEvent } from 'react';
-import { BrowserRouter as Router, Route, Redirect } from 'react-router-dom';
-import SearchBar from './components/@shared/SearchBar/SearchBar';
-import Header from './components/Header/Header';
+import { BrowserRouter as Router, Route, Redirect, Switch } from 'react-router-dom';
 import Tab from './components/Tab/Tab';
 import PATH from './constants/path';
 import HomePage from './pages/HomePage';
-import { MainContainer, Logo } from './styles';
+import { MainContainer } from './styles';
 
 const App = () => {
-  const onSearch = (event: FormEvent<HTMLFormElement>) => {
-    event.preventDefault();
-  };
-
   return (
     <>
       <Router>
-        <Header>
-          <Logo>주절주절</Logo>
-          <SearchBar placeholder="검색어를 입력해주세요" onSubmit={onSearch} />
-        </Header>
-
         <MainContainer>
-          <Route exact path={[PATH.HOME, PATH.ROOT]} component={HomePage} />
-          <Route exact path={[PATH.LOGIN]}>
-            login
-          </Route>
-          <Redirect to={PATH.ROOT} />
+          <Switch>
+            <Route exact path={[PATH.HOME, PATH.ROOT]} component={HomePage} />
+            <Route exact path={[PATH.LOGIN]}>
+              login
+            </Route>
+            <Redirect to={PATH.ROOT} />
+          </Switch>
         </MainContainer>
 
         <Tab></Tab>
