+++ frontend/src/components/Header/MainHeader.styles.ts
@@ -1,0 +1,12 @@
+import styled from '@emotion/styled';
+import { COLOR } from 'src/constants';
+
+const Logo = styled.h1`
+  font-size: 1.4rem;
+  font-weight: 700;
+  color: ${COLOR.WHITE_100};
+  margin-bottom: 1rem;
+  text-align: center;
+`;
+
+export { Logo };
