+++ frontend/src/styles.ts
@@ -10,12 +10,4 @@ const MainContainer = styled.main`
   overflow-y: auto;
 `;
 
-const Logo = styled.h1`
-  font-size: 1.4rem;
-  font-weight: 700;
-  color: ${COLOR.WHITE_100};
-  margin-bottom: 1rem;
-  text-align: center;
-`;
-
-export { MainContainer, Logo };
+export { MainContainer };
